Change end-of-line from DOS to UNIX SL #3607
I had to change the end-of-line[1] to UNIX of xol.py because it was
DOS and the "genpot" command from setup.py (xgettext) was failing with
string with multiple lines. So, it was detecting just the first line
of the multiple lines strings and then the translation string weren't
matching with the source code strings.

[1] https://en.wikipedia.org/wiki/Newline

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>