Export content as HTML SL #3608
Added HTML files to the bundle so this content can be opened with Browse
Activity. The old dita/ditamap files are not removed because they are parsed to
generate the new HTML files.

Besides, the HTML generation uses these dita files to not be too aggressive
with the code in this cycle. We will improve this code in the future and we
will remove completely the dita generation.