Remove unnecessary conditions and extra function
* The reason why the traceback was being observed is that a bool value was being returned
incase the selection from the Gtk.TextView was not over text.
Such a condition is unnecessary, and limits the functionality of the activity.

For example, in the Edit mode, a user cannot drag and drop content from the left side to the right
side. By removing the extra function and the condition, one can do so.

Tested with Gtk version 3.22.30